fixing the controller routes. Dont forget the space in the file name
