--- Models/Departments.js 
@@ -1,12 +1,0 @@
-
-const mongoose = require("../connection");
-const TeamMember = require('./TeamMember.js')
-
-const Schema = mongoose.Schema; 
-
-const departmentSchema = new Schema ({
-    "dptName": String,
-    "member": [TeamMember]
- })
-
- module.exports = mongoose.model("Department", departmentSchema);
