+++ connection.js
@@ -3,7 +3,7 @@ const mongoose = require("mongoose");
 let mongoURL = "";
 
 if (process.env.NODE_ENV === "production") {
-  mongoURL = process.env.DB_URL;
+  mongoURL = "";
 } else {
   mongoURL = "mongodb://127.0.0.1/back-end";
 }
